packages/financial-gateway/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-gateway"
version = "0.0.1"
description = "거래소 및 시뮬레이션과의 통합 인터페이스 게이트웨이"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "생선공장", email = "gatesplan@gmail.com"}
]
license = {text = "MIT"}
dependencies = [
    "financial-assets",
    "financial-simulation",
    "simple-logger",
    "throttled-api[all]",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/gatesplan/python-toolbox"
Repository = "https://github.com/gatesplan/python-toolbox"

[tool.setuptools.packages.find]
include = ["financial_gateway*"]
exclude = ["logs*"]
